working on cgi request

+++ srcs/CgiRequest.cpp
@@ -1,12 +1,30 @@
 #include "CgiRequest.hpp" 
 
+
+/* Execption */
+CgiRequest::cgiException::cgiException(string str) throw() : _str(str) 
+{
+}
+
+CgiRequest::cgiException::~cgiException(void) throw()
+{
+}
+
+const char *
+CgiRequest::cgiException::what(void) const throw()
+{
+	return (_str.c_str());
+}
+
+
+/* CgiRequest */
 CgiRequest::CgiRequest(void)
 {
-	_env[0] = strdup("GATEWAY_INTERFACE=CGI/1.1");
-	_env[1] = strdup("SERVER_PROTOCOL=HTTP/1.1");
-	_env[2] = strdup("SCRIPT_FILENAME=./test.php");
-	_env[3] = strdup("SCRIPT_NAME=test.php");
-	_env[4] = strdup("REDIRECT_STATUS=200");
+	_env[0] = const_cast<char *>("GATEWAY_INTERFACE=CGI/1.1");
+	_env[1] = const_cast<char *>("SERVER_PROTOCOL=HTTP/1.1");
+	_env[2] = const_cast<char *>("SCRIPT_FILENAME=./cgitest/test.php");
+	_env[3] = const_cast<char *>("SCRIPT_NAME=./cgitest/test.php");
+	_env[4] = const_cast<char *>("REDIRECT_STATUS=200");
 	_env[5] = NULL;
 
 	_av[0] = NULL;
@@ -30,35 +48,48 @@ CgiRequest::operator=(CgiRequest const & other)
 }
 
 /* Public method */
-
 void 
 CgiRequest::doRequest(void)
 {
 	int status;
-	int stdOutSave = dup(1);
 	int p[2]; pipe(p);
 	int child = fork();
 	if (child == 0)
 	{
-		dup2(p[1], 1);
-		if (execve("./cgitest/printenv", _av, _env) == -1)
+		dup2(p[1], STDOUT);
+		if (execve("/usr/bin/php-cgi", _av, _env) == -1)
 			exit(EXIT_FAILURE);
 	}
 	else
 	{
-		usleep(100000);
-		waitpid(child, &status, 1);
-		kill(child, SIGKILL);
+		usleep(TIMEOUT);
+		waitpid(child, &status, WNOHANG);
+		if (WEXITSTATUS(status) == EXIT_FAILURE)
+			throw(cgiException("execve fail"));
 
-		ssize_t readReturn = read(p[0], _buffer, BUFFER_SIZE);
-		_buffer[readReturn] = 0;
-		dup2(stdOutSave, 1);
-		cout << _buffer << endl;
+		kill(child, SIGKILL);
+		s_buffer * buffer = NULL;
+		do
+		{
+			buffer = new s_buffer(BUFFER_SIZE);
+			buffer->occupiedSize = read(p[0], buffer->b, static_cast<size_t>(buffer->size));
+			_answer.push(buffer);
+		} while (buffer->occupiedSize == buffer->size);
+		if (buffer->occupiedSize == -1)
+		{
+			deleteQ(_answer);
+			throw(cgiException("read fail"));
+		}
 	}
 }
 
-/* Private method */
+t_bufferQ const & 
+CgiRequest::getAnswer(void) const
+{
+	return (_answer);
+}
 
+/* Private method */
 void
 CgiRequest::_copy(CgiRequest const & other)
 {
